copy fastq files from repo with good ntc if doesn't exist

+++ 1_IlluminaQC.sh
@@ -77,7 +77,17 @@ for variableFile in $(ls *.variables); do
 	# make sample folder
 	mkdir Data/"$sampleId"
 	mv "$variableFile" Data/"$sampleId"
-	mv "$sampleId"_S*.fastq.gz Data/"$sampleId"
+
+	# If the samples fastq files exist, move them to the sample folder
+	if [ -e "$sampleId"_S*.fastq.gz ]; then
+		mv "$sampleId"_S*.fastq.gz Data/"$sampleId"
+
+	# If Data doesn't contain NTC fastq files, copy them from the pipeline dir
+	else
+		 [ ! -e Data/"$sampleId"/"$sampleId"_S*_R1_001.fastq.gz ]; then
+		cp /data/diagnostics/pipelines/"$pipelineName"/"$pipelineName"-"$pipelineVersion"/Data/"$sampleId"/"$sampleId"_S*_R1_001.fastq.gz Data/"$sampleId"
+		cp /data/diagnostics/pipelines/"$pipelineName"/"$pipelineName"-"$pipelineVersion"/Data/"$sampleId"/"$sampleId"_S*_R2_001.fastq.gz Data/"$sampleId"
+	fi
 	
 	# create analysis folders
 	if [[ ! -z ${pipelineVersion-} && ! -z ${pipelineName-} && ! -z ${panel-} && ! -z ${worklistId-} ]]
